Fixed word overlap bug
Continue was used however, as the check is done inside a for loop
instead of the parent while loop, the continue had no effect. It has
been replaced with a flag which does work as intended. A break has also
been included in the foor loop in order to provide some level of
optimization.

Signed-off-by: Matthew Nickson <mnickson@sidingsmedia.com>